软件管理支持本地安装

+++ src/main/core/software/LocalInstall.js
@@ -6,16 +6,13 @@ import CommonInstall from "@/main/core/software/CommonInstall";
 
 export default class LocalInstall {
     static async install(filePath) {
-        let dirName = Path.GetFileNameWithoutExtension(filePath);
-        if (dirName.endsWith(".tar")) {
-            dirName = Path.GetFileNameWithoutExtension(dirName);
-        }
+        const dirName = this.getDirName(filePath);
         const dest = this.getDestPath(dirName);
+        if (!dest) return;
         await CommonInstall.extract(filePath, dest);
         FileUtil.Delete(filePath);
         await CommonInstall.configure(dirName);
     }
-
 
 
     static async installMultiple(files) {
@@ -27,9 +24,14 @@ export default class LocalInstall {
     static getDestPath(dirName) {
         const softList = Software.getList();
         const item = softList.find(item => item.DirName === dirName);
-        if (!item) {
-            throw new Error("离线安装失败，不匹配");
+        return item ? Software.getTypePath(item.Type) : null;
+    };
+
+    static getDirName(filePath) {
+        let dirName = Path.GetFileNameWithoutExtension(filePath);
+        if (dirName.endsWith('.tar')) {
+            dirName = Path.GetFileNameWithoutExtension(dirName);
         }
-        return Software.getTypePath(item.Type);
+        return dirName;
     }
 }
